requirements.txt
streamlit>=1.20
biopython>=1.80
cobra>=0.26.0
python-libsbml>=5.19.0
plotly>=5.10
pandas>=1.5
reportlab>=4.0
lxml>=4.9